api/requirements.txt
fastapi
pydantic
numpy
scikit-learn
openai
python-multipart
Pillow
pypdf
pdfplumber
beautifulsoup4
python-dotenv
